Add files via upload
This program starts with the #include<iosteram>directive to include the input output stream library for using cout and endl. The using namespace std; statement is used to import the entity of the std namespace into the current namespace of the program. The main () function is the entry point of every c++  program .Inside main(), an integer variable number is declared and initialized with the value 678. An if statement checks whether number is greater than 9. If the condition is true, it prints "The number is greater than 9." using cout . If the condition is false, the else block prints "The number is 9 or less". The program ends by returning 0,Which indicates that the program has executed successfully .