Add files via upload
The program starts by including the<iostream>header, which is necessary for input and output operations . using namespace std; is used to avoid writing std:: before standard library functions like cout and cin .The main function is the entry point of the program. Inside the main function, an integer variable number is declared. The program prompts the user to enter an integer using cout .The user input is captured and stored in the number variable using cin . After receiving the input , the program displays the entered number using cout .The program outputs the message "You entered" followed by the entered number. The return 0; statement indicates that the program executed successfully.