Add files via upload
This C++ code calculates and displays the sum of all natural numbers from 1 to a user-specified positive integer n. The program first prompts the user to input a positive integer, which is stored in the variable n. It then initializes a variable sum to 0, which will accumulate the sum of the numbers. A for loop iterates from 1 to n, incrementing the sum by each number in the range. Finally, the program outputs the calculated sum to the console.